requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.7.2
scikit-image==0.21.0
tensorflow==2.13.0
scipy==1.11.1
imutils==0.5.4
argparse==1.4.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1


